Rename custom-field.css to abstract-custom-field.css
