+++ src/components/__internal__/abstract-custom-field.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import Icon from '../icon/icon.jsx';
-import styles from './custom-field.css';
+import styles from './abstract-custom-field.css';
 import cautionSvg from '../../svgs/icon-caution-fill.svg';
 import FormControl from '../form-control/form-control.jsx';
