+++ src/components/custom-field-input-multiple-choice/custom-field-input-multiple-choice.css
@@ -1,6 +1,6 @@
 .container {
   composes: tag-list from '../tag-list/tag-list.css';
-  composes: input-container from '../__internal__/custom-field.css';
+  composes: input-container from '../__internal__/abstract-custom-field.css';
   height: auto;
   min-height: 32px;
   padding-top: 0;
@@ -10,14 +10,14 @@
 
 .read-write-container {
   composes: container;
-  composes: input from '../__internal__/custom-field.css';
+  composes: input from '../__internal__/abstract-custom-field.css';
 }
 
 .read-only-container {
   composes: container;
-  composes: read-only-input from '../__internal__/custom-field.css';
+  composes: read-only-input from '../__internal__/abstract-custom-field.css';
 }
 
 .input-icon {
-  composes: input-icon from '../__internal__/custom-field.css';
+  composes: input-icon from '../__internal__/abstract-custom-field.css';
 }
